requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
tqdm
transformers
opencv-python
datasets
accelerate